refactor(Dockerfile): use Debian base images
This allows us to use wheels when installing Python dependencies, which
seems to save a _lot_ of time on ARM64, since it avoids compiling
LAPACK.

The resulting image is roughly twice the size (~84MB compressed) but the
build time is significantly lower (~1600s vs ~3600s for the ARM64
image).

Note that this change also changes the pathfinder binary to be a
dynamically linked executable, and this means that we have to make sure
that the rust-builder and the runner images are compatible. We're
explicitly using Rust and Python images that are based on Debian
bullseye to avoid incompatibilities.